requirements.txt
streamlit
pandas
matplotlib
folium
google-cloud-bigquery
